implement `prepare` init container
The `prepare` container has the same config as the `upgrade` container.
If configured it will be the first container to run in the pod. This enables
the use-case where a plan to be applied to worker nodes may want to wait on
another plan being applied to master nodes before incurring drain/cordon on
the workers. There are likely man other legitimate use-cases.