(#59) Do not overwrite container image tag if provided
The current behavior is to force the upgrade container image tag
to be overwritten by status.lastVersion.
While this works for some workflows, it prevents workflows where the
upgrade is done in a non versioned image (or that always requires
latest for instance), because it becomes impossible to mention a
given specific image tag (and especially when using channel instead
of a specific Plan version).

This change implements the following behavior:
* if the image is not specified with a tag, the status.lastVersion
is used, as before.
* if the image specifies a tag, then this tag is used