latest tag for prepare and upgrade containers
In reviewing #60 I realized that the `WithImageTag` function was buggy
as well as the tests for it. So, in the spirit of the changes presented
in #60 (requested in #59) I have replaced it with `WithLatestTag` which
will not override a tag if an image reference already has one. This
means that SUC interprets a tag-less image ref specified for `prepare`
or `upgrade` containers to have the plan's resolved version (aka
`.status.latestVersion`) instead of `latest`.

Fixes #41
Fixes #59