Make pkg/apis a go module (#102)
* Make pkg/apis a go module

system-upgrade-controller depends on k8s.io/kubernetes which makes it very
difficult to include as a dependency from another project. The main reason
a project would want to have a go dependency on system-upgrade-controller
would be for the types. This makes the apis package a first class module
that one can depend on without having to depend on k8s.io/kubernetes

* Swith to vendorless