Allow adding sections of plan to hash via annotation (#155)
Previously, only the latestVersion, serviceAccountName, and secrets were
used in the hash value to track updates. This meant that updating other
parts of the plan (like environment variables) would not automatically
trigger an update.

After this change, if the plan has an annotation
(upgrade.cattle.io/digest) with a value of a comma-delimited string of
pieces of the plan, they will be included in the hash and tracked for
updates.