By default restart pods on failure
Prior to this is a pod failed for any reason we would not recover for
maybe 15 minutes.  By setting the RestartPolicy to OnFailure the jobs
will be more resilient to transient failure.  The big issue we are trying
to address is when the kubelet fails to launch a pod because it
can't find the host-info volume, which is transient.