Add default fields in Job (#173)
With the new version of wrangler, default fields are added. This poses
an issue when Apply is called again because these default fields won't
exist on the next call. This causes wrangler to think that something has
changed.

In this fix, default fields are added to the Job to prevent this issue.