fix(kustomize): Deploy latest system-upgrade-controller
Before this patch always version v0.8.0 would be deployed when using the
kustomize command from the README. This patch adjust the kustomization
to not overwrite the version.

The build process is unaffected from this, since during the build script
the kustomize file is adjusted to adjust the version to the current
release. This means the build manifests for release will still have a
pinned version like before.

Reference:
https://github.com/rancher/system-upgrade-controller/blob/4eacc2dabbde943ee59dc7819ea0a04909b05b0f/scripts/package-controller#L36