ci(drone): Add push step for latest manifest
This patch adds a push step for a latest image. Various CI steps
actually build a latest image but it's never pushed. This patch adds a
step in the release pipeline to also push the release as latest to
Docker Hub.

This should resolve issues people experience when using plain manifests
from the repository which still references a latest image, which
currently doesn't exist.

References:
https://github.com/rancher/system-upgrade-controller/issues/302