Upgrade dependencies
Some Golang CVEs get flagged in the suc binary as it was built with an
older golang version. Use go 1.22.6
Updating other go module deps.
SLES update is required for libc compatiblity with latest ginkgo
version.

Update golangci-lint. Most gingko and gomega changes are to fix linting
failures.
Update sonobuoy

docker bridge address is added to tls-san to avoid cert errors when
using the client to connect the KUBEHOST.

Drop deprecated version field in the docker compose manifests.

It is recommended that the next release be v0.15.0 as 0.14.0 was already
published and deleted in the past and gosumdb already has entries.